requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pytest
python-dotenv
aiomysql